input: joystick: Fix buffer data parsing
Don't try to access buffer data of a channel by its scan index. Instead,
calculate its offset in the buffer.

This is necessary, as the scan index of a channel does not represent its
position in a buffer - the buffer will contain data for enabled channels
only, affecting data offsets and alignment.

While at it, also fix minor style issue in probe.

Reported-by: Chris Morgan <macromorgan@hotmail.com>
[pcercuei: Fix race vs. IRQ handler]
Signed-off-by: Paul Cercueil <paul@crapouillou.net>